+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v2.0.22.3
+- Show magic popup when only the video has a magic assigned
+
 ## v2.0.22.2
 - Disable autoplay for notification popup
 - Reduce scroll to replacement of walkthrough video